backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
aiofiles==24.1.0
pypdf==4.3.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic[email]==2.9.2